Update chức năng thông báo, chia component cho trang detail
