+++ src/components/admin/ComponentCon/Header/NavIcon.tsx
@@ -1,5 +1,5 @@
 import { Link } from 'react-router-dom';
-import { HomeOutlined, MenuOutlined } from '@ant-design/icons';
+import { HomeOutlined, MenuOutlined, SettingOutlined } from '@ant-design/icons';
 import { Tooltip } from 'antd';
 
 export default function NavIcons() {
@@ -15,11 +15,11 @@ export default function NavIcons() {
         </Link>
       </div>
 
-      {/* <Link to="#">
+      <Link to="/admin/seting">
         <button className="p-2 rounded-full hover:bg-gray-100 transition-colors">
           <SettingOutlined className="text-gray-600" style={{ fontSize: 20 }} />
         </button>
-      </Link> */}
+      </Link>
 
       <button className="md:hidden p-2 rounded-full hover:bg-gray-100 transition-colors">
         <MenuOutlined className="text-gray-600" style={{ fontSize: 22 }} />
